requirements.txt
can-isotp==1.8
msgpack==1.0.4
packaging==21.3
pyparsing==3.0.9
python-can==4.1.0
typing-extensions==4.4.0
wrapt==1.14.1
